fix(ci): runtime file-length cap + integration-test lint acknowledgments
- Extract the driver loop (driver_loop, registry-event processing, oplog
  tailing, config-mtime backstop) from runtime/mod.rs into the sibling
  module runtime/driver.rs — mod.rs was 532 lines, over the 500-line
  structure cap.
- Acknowledge the new minisign-verify dependency (`use minisign_verify
  as _;`) in the six integration-test targets, per the per-target
  unused-crate-dependencies lint (openapi/main.rs already had it).
- Drop unnecessary `crate::services::` qualifications in the
  rest/config/update.rs tests (unused-qualifications is deny).
