+++ crates/cp-orchestrator/tests/registry_channel.rs
@@ -33,6 +33,7 @@ use cp_base as _;
 use cp_vault as _;
 use csv as _;
 use dotenvy as _;
+use minisign_verify as _;
 use nix as _;
 use notify as _;
 use openssl as _;
