+++ crates/cp-orchestrator/tests/supervisor_lifecycle.rs
@@ -29,6 +29,7 @@ use cp_oplog as _;
 use cp_vault as _;
 use csv as _;
 use dotenvy as _;
+use minisign_verify as _;
 use notify as _;
 use openssl as _;
 use portable_pty as _;
